+++ packages/apis/uniswapv2/src/utils/Currency.ts
@@ -1,7 +1,0 @@
-import { Currency } from "../query/w3";
-
-export const ETHER: Currency = {
-  decimals: 18,
-  name: "Ether",
-  symbol: "ETH",
-};