+++ packages/schema/parse/src/header.ts
@@ -20,5 +20,11 @@ directive @imported(
 directive @imports(
   types: [String!]!
 ) on OBJECT
+
+directive @capability(
+  type: String!
+  uri: String!
+  namespace: String!
+) repeatable on OBJECT
 ### Web3API Header END ###
 `;