+++ packages/test-cases/cases/bind/sanity/output/wasm-as/query/TestImport/index.ts
@@ -1,0 +1,11 @@
+import {
+  w3_getImplementations
+} from "@web3api/wasm-as";
+
+export class TestImport {
+  static uri: string = "testimport.uri.eth"
+
+  public static getImplementations(): string[] {
+    return w3_getImplementations(this.uri);
+  }
+}