+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Web3API 0.0.1-prealpha.50
+## Features
+* Getting the implementations of an interface is now supported from within Wasm.
+* `@web3api/tracing-js`: Added a class method decorator for tracing.
+
 # Web3API 0.0.1-prealpha.49
 ## Features
 * `@web3api/fs-plugin-js`: Added a "File System" plugin, which implements the `uri-resolver` interface, enabling users to load Web3API packages from their local filesystem. For example, a user could specify the URI `/fs/path/to/package/directory`.